LayoutTests/imported/w3c/web-platform-tests/feature-policy/reporting/serial-reporting.https-expected.txt


FAIL requestPort in serial reporting mode assert_equals: expected (number) 18 but got (undefined) undefined
FAIL getPorts in serial reporting mode assert_equals: expected (number) 18 but got (undefined) undefined
